// backend/src/dto/calculate-fso.dto.ts
import { EnumModel, EnumAirTurbulence, EnumCalcMethod } from '../logic/fso-math.logic';

export class CalculateFsoDto {
  model: EnumModel;
  airTurbulence: EnumAirTurbulence;
  adtUtlmMethod: EnumCalcMethod;
  
  vykon: number;
  jednotkaVykonu: number;
  citlivost: number;
  priemer: number;
  smerovost: number;
  vlnovaDlzka: number;
  
  vzdialenost: number;
  jednotkaVzdialenosti: number;
  
  viditelnost: number;
  
  dazd: number;
  suchySneh: number;
  mokrySneh: number;
  
  aditivnyUtlm: number;
}